pyproject.toml
[tool.poetry]
name = "wave"
version = "0.1.0"
description = ""
authors = ["J-M-J-R <52387908+J-M-J-R@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
cattrs = "^1.1.2"
attrs = "^20.3.0"
h2o-wave = "^0.10.0"
altair = "^4.1.0"
minizinc = "^0.4.2"
pendulum = "^2.1.2"
logzero = "^1.6.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
mypy = "^0.790"
pytest-asyncio = "^0.14.0"
pytest-watch = "^4.2.0"

[tool.pytest.ini_options]
addopts = "-ra -q -s"
markers = [ "watch" ]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
